requirements.txt
flask==2.0.1
google-generativeai==0.3.2
python-dotenv==0.19.0
werkzeug==2.0.1 